Adding 'translation-redirect' Middleware
