+++ config/localization.php
@@ -23,6 +23,7 @@ return [
             'localization-cookie-redirect'  => false,
             'localization-redirect'         => true,
             'localized-routes'              => true,
+            'translation-redirect'          => true,
         ],
     ],
