+++ src/Providers/RoutingServiceProvider.php
@@ -4,6 +4,7 @@ use Arcanedev\Localization\Middleware\LocaleCookieRedirect;
 use Arcanedev\Localization\Middleware\LocaleSessionRedirect;
 use Arcanedev\Localization\Middleware\LocalizationRedirect;
 use Arcanedev\Localization\Middleware\LocalizationRoutes;
+use Arcanedev\Localization\Middleware\TranslationRedirect;
 use Illuminate\Routing\RoutingServiceProvider as ServiceProvider;
 use Arcanedev\Localization\Routing\Router;
 
@@ -64,10 +65,11 @@ class RoutingServiceProvider extends ServiceProvider
         /** @var Router $router */
         $router      = $this->app['router'];
         $middlewares = [
-            'localized-routes'              => LocalizationRoutes::class,
             'localization-session-redirect' => LocaleSessionRedirect::class,
             'localization-cookie-redirect'  => LocaleCookieRedirect::class,
             'localization-redirect'         => LocalizationRedirect::class,
+            'localized-routes'              => LocalizationRoutes::class,
+            'translation-redirect'          => TranslationRedirect::class,
         ];
 
         foreach ($middlewares as $name => $class) {
